requirements.txt
gradio
transformers
torch
Pillow
fugashi
protobuf
unidic_lite
manga-ocr
fastapi